apply DRY
