+++ src/test/AddFieldOrGroupButton.test.tsx
@@ -1,4 +1,4 @@
-import { act, render, screen } from '@testing-library/react'
+import { render, screen } from '@testing-library/react'
 
 import AddFieldOrGroupButton from '../components/customSearch/AddFieldOrGroupButton'
